Create Tokenomics README.md (#32)
* Create Tokenomics README.md

* Create placeholder.txt

* Add files via upload

* Update ETH_Gas_Data_Infura.py

Added async/parallel processing updates and error handling to all 3 scripts, making them 50-100x more efficient. Chainstack is currently the main version of the script for block data scraping, with the other 2 (Infura and Etherscan) are additional tools for use with chains besides Ethereum or as backups.

* Update ETH_Gas_Data_Etherscan.py

Same as previous update

* Update ETH_Gas_Data_Chainstack.py

Same updates as the last two scripts, this is the current main version between the three. Aync/concurrency and error handling updates, as well as additional columns to assist in GNOT fee simulation.

* Update README.md

Added a description of the revised/improved script functionality